charts_web/mybackend/requirements.txt
accelerate==0.33.0
ace_tools==0.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.0.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.42.0
black==24.8.0
bleach==6.1.0
blinker==1.8.2
boltons==21.0.0
bracex==2.5
Brotli==1.1.0
cachelib==0.13.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
click-option-group==0.5.6
codeshield==1.0.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.5
datasets==2.21.0
DateTime==5.5
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
enum34==1.1.10
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.0
face==22.0.0
fastapi==0.112.1
fastjsonschema==2.18.1
ffmpy==0.4.0
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
Flask-Session==0.8.0
fonttools==4.44.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
glom==22.1.0
googleapis-common-protos==1.63.2
gradio==4.41.0
gradio_client==1.3.0
groq==0.9.0
h11==0.14.0
html5lib==1.1
http-client==0.1.22
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.4
importlib_metadata==7.1.0
importlib_resources==6.4.3
inflate64==1.0.0
ipykernel==6.25.2
ipympl==0.9.3
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
jiter==0.5.0
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
langchain==0.2.14
langchain-community==0.0.36
langchain-core==0.2.34
langchain-text-splitters==0.2.2
langsmith==0.1.101
LinearRegression==0.0.1
llama-recipes==0.0.3
loralib==0.1.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.2
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
msgspec==0.18.6
multidict==6.0.5
multiprocess==0.70.16
multivolumefile==0.2.3
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.3
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.0
openai==1.39.0
openpyxl==3.1.2
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
optimum==1.21.4
orjson==3.10.3
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.12.1
peewee==3.17.6
peft==0.12.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
praw==7.7.1
prawcore==2.4.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==4.25.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py7zr==0.22.0
pyarrow==17.0.0
pybcj==1.0.2
pycparser==2.21
pycryptodomex==3.20.0
pycurl==7.45.2
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.1.1
pyppmd==1.1.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
pyzstd==0.16.1
qtconsole==5.5.0
QtPy==2.4.1
referencing==0.30.2
regex==2024.5.15
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.12.0
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
ruff==0.6.1
safetensors==0.4.4
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.2
semantic-version==2.10.0
semgrep==1.85.0
Send2Trash==1.8.2
sentencepiece==0.2.0
setuptools==68.2.2
shellingham==1.5.4
six==1.16.0
sk==0.0.1
sklearn==0.0.post11
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.38.2
sympy==1.13.2
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
terminado==0.17.1
textblob==0.18.0.post0
texttable==1.7.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenize-rt==6.0.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
torch==2.4.0
tornado==6.3.3
tqdm==4.66.4
traitlets==5.11.2
transformers==4.43.4
typer==0.12.4
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
update-checker==0.18.0
uri-template==1.3.0
urllib3==2.0.6
uvicorn==0.30.6
wcmatch==8.5.2
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.3
wget==3.2
widgetsnbextension==4.0.9
wrapt==1.16.0
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.9.4
zipp==3.20.0
zope.interface==7.0.3
